requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
altgraph==0.17.3
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
bcrypt==4.0.1
cachetools==5.3.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
confluent-kafka==2.1.1
crypto==1.4.1
Crypto.Cipher==1
Crypto20373215155==2022.6.30
cryptography==39.0.1
Deprecated==1.2.13
docopt==0.6.2
elastic-transport==8.4.0
elasticsearch==8.6.2
Flask==2.2.3
Flask-Limiter==3.3.1
frozenlist==1.3.3
idna==3.4
importlib-resources==5.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
kafka==1.3.5
kafka-python==2.0.2
limits==3.5.0
logging==0.4.9.6
markdown-it-py==2.2.0
MarkupSafe==2.1.2
mdurl==0.1.2
multidict==6.0.4
mysql-connector-python==8.0.33
Naked==0.1.32
numpy==1.24.2
openai==0.27.2
opencv-python==4.7.0.68
ordered-set==4.1.0
packaging==23.1
Padding==0.5
panda==0.3.1
paramiko==3.1.0
pefile==2023.2.7
pipreqs==0.4.13
ply==3.11
prometheus-client==0.17.0
protobuf==3.20.3
psutil==5.9.4
py2exe==0.13.0.0
pyaes==1.6.1
pyasn1==0.4.8
pycodestyle==2.10.0
pycparser==2.21
pycryptodome==3.18.0
pycryptodomex==3.17
pygame==2.4.0
Pygments==2.15.1
pyinstaller==5.8.0
pyinstaller-hooks-contrib==2023.0
PyNaCl==1.5.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.11.1
pysmi==0.3.4
pysnmp==4.4.12
python-ffmpeg-video-streaming==0.1.15
python-vlc==3.0.18121
pytube==15.0.0
pywin32-ctypes==0.2.0
PyYAML==6.0
requests==2.28.2
rich==13.3.5
shellescape==3.8.1
tomli==2.0.1
tqdm==4.65.0
typing_extensions==4.5.0
urllib3==1.26.14
Werkzeug==2.2.3
wrapt==1.15.0
yarg==0.1.9
yarl==1.8.2
